requirements.txt
streamlit
sentence-transformers
faiss-cpu
numpy
PyPDF2
scikit-learn